LayoutTests/imported/w3c/web-platform-tests/fullscreen/api/element-request-fullscreen-active-document-expected.txt


PASS Element#requestFullscreen() when the document is not the active document
